Clean up with Room, MVVM and gitignore
