+++ .gitignore
@@ -1,7 +1,7 @@
 *.iml
 .gradle
 /local.properties
-.idea/*
+.idea/
 !.idea/codeStyles/
 !.idea/inspectionProfiles/
 .DS_Store
