+++ app/src/main/java/com/iyanuoluwa/imovie/MovieApplication.kt
@@ -1,6 +1,8 @@
-package com.iyanuoluwa.imovie.data.local
+package com.iyanuoluwa.imovie
 
 import android.app.Application
+import com.iyanuoluwa.imovie.data.local.MovieDatabase
+import com.iyanuoluwa.imovie.ui.main.MovieRepository
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
 
@@ -9,5 +11,5 @@ class MovieApplication : Application() {
     private val applicationScope = CoroutineScope(SupervisorJob())
 
     private val database by lazy { MovieDatabase.getDatabase(this) }
-    private val repository by lazy { MovieRepository(database.movieDao()) }
+    val repository by lazy { MovieRepository(database.movieDao()) }
 }
