+++ app/src/main/java/com/iyanuoluwa/imovie/data/local/MovieDatabase.kt
@@ -1,9 +1,14 @@
 package com.iyanuoluwa.imovie.data.local
 
 import android.content.Context
+import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import androidx.room.TypeConverters
+import com.iyanuoluwa.imovie.data.model.Result
 
+@Database(entities = [Result::class], version = 1, exportSchema = false)
+@TypeConverters(Converters::class)
 abstract class MovieDatabase : RoomDatabase() {
 
     abstract fun movieDao() : MovieDao
@@ -13,15 +18,12 @@ abstract class MovieDatabase : RoomDatabase() {
         private var INSTANCE : MovieDatabase? = null
 
         fun getDatabase(context: Context) : MovieDatabase {
-
-            return INSTANCE ?: kotlin.synchronized(this) {
-
+            return INSTANCE ?: synchronized(this) {
                 val instance = Room.databaseBuilder(
                     context.applicationContext,
                     MovieDatabase::class.java,
-                    "iMovie"
-                )
-                    .build()
+                    "iMovie.db"
+                ).build()
                 INSTANCE = instance
                 instance
             }
