+++ app/src/main/java/com/iyanuoluwa/imovie/data/model/Result.kt
@@ -1,20 +1,18 @@
 package com.iyanuoluwa.imovie.data.model
 
-
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import com.google.gson.annotations.SerializedName
 
 @Entity(tableName = "movies_table")
 data class Result(
-    @PrimaryKey(autoGenerate = true)
-    val ids : Int,
+    @PrimaryKey
+    val id: Int,
     val adult: Boolean,
     @SerializedName("backdrop_path")
     val backdropPath: String,
     @SerializedName("genre_ids")
     val genreIds: List<Int>,
-    val id: Int,
     @SerializedName("original_language")
     val originalLanguage: String,
     @SerializedName("original_title")
