+++ app/src/main/java/com/iyanuoluwa/imovie/ui/main/MovieRepository.kt
@@ -1,6 +1,7 @@
-package com.iyanuoluwa.imovie.data.local
+package com.iyanuoluwa.imovie.ui.main
 
 import androidx.annotation.WorkerThread
+import com.iyanuoluwa.imovie.data.local.MovieDao
 import com.iyanuoluwa.imovie.data.model.Result
 import kotlinx.coroutines.flow.Flow
 
@@ -8,9 +9,9 @@ class MovieRepository(private val movieDao : MovieDao) {
 
     val allMovies : Flow<List<Result>> = movieDao.getMovieDatabase()
 
-    @Suppress("RedundantSuspendModifier")
     @WorkerThread
-    suspend fun insert(movie : Result) {
-        movieDao.insert(movie)
+    suspend fun insertMovies(movies : List<Result>) {
+        movieDao.insertAllMovies(movies)
     }
+
 }
