+++ app/src/main/java/com/iyanuoluwa/imovie/ui/main/fragments/TopRated.kt
@@ -1,4 +1,4 @@
-package com.iyanuoluwa.imovie.ui.main
+package com.iyanuoluwa.imovie.ui.main.fragments
 
 import android.os.Bundle
 import android.util.Log
@@ -13,7 +13,9 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.iyanuoluwa.imovie.R
 import com.iyanuoluwa.imovie.data.model.MovieJson
-import com.iyanuoluwa.imovie.data.remote.ApiPopular
+import com.iyanuoluwa.imovie.data.remote.ApiTopRated
+import com.iyanuoluwa.imovie.ui.main.MainActivity
+import com.iyanuoluwa.imovie.ui.main.MovieAdapter
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
@@ -21,9 +23,9 @@ import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
 
-class Popular : Fragment() {
+class TopRated : Fragment() {
 
-    private var popularRecyclerView: RecyclerView? =null
+    private var topRatedRecyclerView: RecyclerView? =null
     private var gridLayoutManager: GridLayoutManager? = null
     private var titleList = mutableListOf<String>()
     private var imageList = mutableListOf<String>()
@@ -38,15 +40,15 @@ class Popular : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
         // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_popular, container, false)
-        popularRecyclerView = view.findViewById(R.id.popular_recycler_view)
+        val view = inflater.inflate(R.layout.fragment_top_rated, container, false)
+        topRatedRecyclerView = view.findViewById(R.id.top_rated_recycler_view)
         gridLayoutManager = GridLayoutManager(context, 3, LinearLayoutManager.VERTICAL, false)
-        nestedScrollView = view.findViewById(R.id.scroll_view_popular)
-        progressBar = view.findViewById(R.id.progress_bar_popular)
+        nestedScrollView = view.findViewById(R.id.scroll_view_topRated)
+        progressBar = view.findViewById(R.id.progress_bar_topRated)
         return view
     }
 
-    private fun addToList(title : String, image: String, plot: String, id: Int) {
+    private fun addToList(title: String, image: String, plot: String, id: Int) {
         titleList.add(title)
         imageList.add(image)
         plotList.add(plot)
@@ -54,39 +56,39 @@ class Popular : Fragment() {
     }
 
     private fun setUpRecyclerView() {
-        popularRecyclerView?.layoutManager = gridLayoutManager
-        popularRecyclerView?.adapter = context?.let { MovieAdapter(it, titleList, imageList, plotList, ids) }
+        topRatedRecyclerView?.layoutManager = gridLayoutManager
+        topRatedRecyclerView?.adapter = context?.let { MovieAdapter(it, titleList, imageList, plotList, ids) }
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        getPopularMovies(page, limit)
+        getTopRatedMovies(page, limit)
         nestedScrollView?.setOnScrollChangeListener(object : NestedScrollView.OnScrollChangeListener {
             override fun onScrollChange(v: NestedScrollView?, scrollX: Int, scrollY: Int, oldScrollX: Int, oldScrollY: Int) {
                 if (scrollY == (v?.getChildAt(0)?.measuredHeight)?.minus(v?.measuredHeight!!)) {
                     page++
                     progressBar?.visibility = View.VISIBLE
-                    getPopularMovies(page, limit)
+                    getTopRatedMovies(page, limit)
                 }
             }
         })
     }
 
-    private fun getPopularMovies(page: Int, limit: Int) {
-        val apiPopular = Retrofit.Builder()
-                .baseUrl(BASE_URL)
+    private fun getTopRatedMovies(page: Int, limit: Int) {
+        val apiTopRated = Retrofit.Builder()
+                .baseUrl(MainActivity.BASE_URL)
                 .addConverterFactory(GsonConverterFactory.create())
                 .build()
-                .create(ApiPopular::class.java)
+                .create(ApiTopRated::class.java)
 
-        val data = apiPopular.getPopularMovies(page, limit)
+        val data = apiTopRated.getTopRatedMovies(page, limit)
         data.enqueue(object : Callback<MovieJson?> {
             override fun onResponse(call: Call<MovieJson?>, response: Response<MovieJson?>) {
                 if (response.isSuccessful && response.body() != null) {
                     progressBar?.visibility = View.GONE
                     val responseBody = response.body()!!
                     for (movies in responseBody.results) {
-                        Log.i("PopularFragment", "Result = $movies")
+                        Log.i("TopRatedFragment", "Result = $movies")
                         addToList(movies.title, "https://image.tmdb.org/t/p/w500${movies.posterPath}", movies.overview, movies.id)
                     }
                     setUpRecyclerView()
@@ -94,7 +96,7 @@ class Popular : Fragment() {
             }
 
             override fun onFailure(call: Call<MovieJson?>, t: Throwable) {
-                Log.e("PopularFragment", t.toString())
+                Log.e("TopRatedFragment", t.toString())
             }
         })
     }
