+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,7 @@
     android:background="@color/black"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.details.MainActivity">
+    tools:context=".ui.main.MainActivity">
