+++ app/src/main/res/layout/fragment_popular.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.main.Popular">
+    tools:context=".ui.main.fragments.Popular">
 
     <TextView
         android:id="@+id/popular"
