+++ app/src/main/res/layout/fragment_top_rated.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.main.TopRated">
+    tools:context=".ui.main.fragments.TopRated">
 
     <TextView
         android:id="@+id/top_rated"
